feat: reuse getNotificationPermissionsStatus
