+++ src/lib/Scenes/SavedSearchAlert/SavedSearchAlertForm.tsx
@@ -1,9 +1,9 @@
 import { FormikProvider, useFormik } from "formik"
 import { getSearchCriteriaFromFilters } from "lib/Components/ArtworkFilter/SavedSearch/searchCriteriaHelpers"
 import { LegacyNativeModules } from "lib/NativeModules/LegacyNativeModules"
+import { getNotificationPermissionsStatus, PushAuthorizationStatus } from "lib/utils/PushNotification"
 import React from "react"
 import { Alert, AlertButton, Linking, Platform } from "react-native"
-import { getNotificationPermissionsStatus, PushAuthorizationStatus } from "../MyProfile/MyProfilePushNotifications"
 import { Form } from "./Components/Form"
 import { extractPills, getNamePlaceholder } from "./helpers"
 import { createSavedSearchAlert } from "./mutations/createSavedSearchAlert"
